backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pydantic==1.10.7
sqlalchemy==2.0.9
python-jose==3.3.0
passlib==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.0
python-multipart==0.0.6
transformers==4.30.2
torch==2.0.1
ollama==0.1.9